Move mocked data to shared module
Move mocked data to shared module
